Fix failed refactoring
