+++ src/main/java/org/jsonstat/v2/DimensionNotFoundException.java
@@ -1,4 +1,4 @@
-package org.jsonstat;
+package org.jsonstat.v2;
 
 /**
  * Created by hadrien on 15/06/16.
