+++ src/test/resources/org/jsonstat/v2/collection.json
