+++ src/test/resources/org/jsonstat/v2/galicia.json
