+++ src/test/resources/org/jsonstat/v2/us-labor.json
